get.go for handlers
